refactor: 몇몇 코드 리팩토링
