+++ src/app/home/challenge/[challengeId]/_componenets/CategoryPieChart/CategoryPieChart.tsx
@@ -3,7 +3,6 @@ import { CategoryPieChartProps } from "./CategoryPieChart.type";
 import useCategoryPieChart from "./CategoryPieChart.hook";
 
 export default function CategoryPieChart({ expenses }: CategoryPieChartProps) {
-    console.log(expenses);
     const { expensesByCategory } = useCategoryPieChart({ expenses });
 
     return (
