+++ src/app/home/challenge/[challengeId]/_componenets/LinkToEditPage/LinkToEditPage.type.ts
@@ -1,0 +1,3 @@
+export interface LinkToEditPageProps {
+    challengeId: string | undefined;
+}
