+++ src/app/home/challenge/[challengeId]/_componenets/Main/Main.hook.tsx
@@ -1,15 +1,22 @@
 import { getChallengeById } from "@/apis/services/challenge";
 import { getExpensesByChallengeDuration } from "@/apis/services/expense";
+import useModalStore from "@/stores/modalStore";
 import { ChallengeResponse } from "@/types/challenge";
 import { ExpenseData } from "@/types/expense";
 import { useParams, useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
+import DeleteChallengeModal from "../DeleteChallengeModal/DeleteChallengeModal";
 
 export default function useMain() {
     const { challengeId } = useParams();
+    const { setIsModalOpen } = useModalStore();
 
     const [challenge, setChallenge] = useState<ChallengeResponse>();
     const [expenses, setExpenses] = useState<ExpenseData[]>([]);
+
+    const handleDeleteChallengeModalOpen = () => {
+        setIsModalOpen(<DeleteChallengeModal />);
+    };
 
     useEffect(() => {
         getChallengeById({
@@ -39,5 +46,6 @@ export default function useMain() {
     return {
         challenge,
         expenses,
+        handleDeleteChallengeModalOpen,
     };
 }
