+++ src/app/home/challenge/[challengeId]/_componenets/Main/Main.tsx
@@ -4,9 +4,11 @@ import ExpenseOfChallengeContainer from "../ExpenseOfChallengeContainer/ExpenseO
 import DailyExpenseBarChart from "../DailyExpenseBarChart/DailyExpenseBarChart";
 import ChallengeInfo from "../ChallengeInfo/ChallengeInfo";
 import useMain from "./Main.hook";
+import LinkToEditPage from "../LinkToEditPage/LinkToEditPage";
+import DangerousButton from "@/components/button/DangerousButton";
 
 export default function Main() {
-    const { challenge, expenses } = useMain();
+    const { challenge, expenses, handleDeleteChallengeModalOpen } = useMain();
 
     return (
         <main className="flex flex-col gap-8">
@@ -30,6 +32,23 @@ export default function Main() {
                 {expenses && (
                     <ExpenseOfChallengeContainer expenses={expenses} />
                 )}
+            </section>
+
+            <section className="flex flex-col gap-4">
+                <h3 className="text-xl font-bold">챌린지 관리</h3>
+                <div className="flex flex-col gap-2">
+                    <LinkToEditPage challengeId={challenge?.id} />
+
+                    <DangerousButton
+                        type="button"
+                        text="삭제"
+                        rounded="rounded-md"
+                        px="px-2"
+                        py="py-1"
+                        width="w-auto"
+                        onClick={handleDeleteChallengeModalOpen}
+                    />
+                </div>
             </section>
         </main>
     );
